--- src/by/it/VeronicaDemesh/lesson02/B_Sheduler.java
@@ -1,71 +1,0 @@
-package by.it.VeronicaDemesh.lesson02;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.List;
-
-public class B_Sheduler {
-    //событие у аудитории(два поля: начало и конец)
-    static class Event {
-        int start;
-        int stop;
-
-        Event(int start, int stop) {
-            this.start = start;
-            this.stop = stop;
-        }
-
-        @Override
-        public String toString() {
-            return "("+ start +":" + stop + ")";
-        }
-    }
-
-    public static void main(String[] args) {
-        B_Sheduler instance = new B_Sheduler();
-        Event[] events = {  new Event(0, 3),  new Event(0, 1), new Event(1, 2), new Event(3, 5),
-                            new Event(1, 3),  new Event(1, 3), new Event(1, 3), new Event(3, 6),
-                            new Event(2, 7),  new Event(2, 3), new Event(2, 7), new Event(7, 9),
-                            new Event(3, 5),  new Event(2, 4), new Event(2, 3), new Event(3, 7),
-                            new Event(4, 5),  new Event(6, 7), new Event(6, 9), new Event(7, 9),
-                            new Event(8, 9),  new Event(4, 6), new Event(8, 10), new Event(7, 10)
-                          };
-
-        List<Event> starts = instance.calcStartTimes(events,0,10);  //рассчитаем оптимальное заполнение аудитории
-        System.out.println(starts);                                 //покажем рассчитанный график занятий
-    }
-
-    List<Event> calcStartTimes(Event[] events, int from, int to) {
-        //events - события которые нужно распределить в аудитории
-        //в период [from, int] (включительно).
-        //оптимизация проводится по наибольшему числу непересекающихся событий.
-        //начало и конец событий могут совпадать.
-        List<Event> result;
-        result = new ArrayList<>();
-        //ваше решение.
-        Arrays.sort(events, new Comparator<Event>() {
-            @Override
-            public int compare(Event o1, Event o2) {
-                int stop1 = o1.stop;
-                int stop2 = o2.stop;
-                return stop1 - stop2;
-            }
-        });
-        int i = 0;
-        int stop;
-        while(i < events.length){
-           stop = events[i].stop;
-           result.add(events[i]);
-           while(i < events.length){
-                if(events[i].start < stop){
-                    i++;
-                }else{
-                    break;
-                }
-            }
-
-        }
-        return result;                        //вернем итог
-    }
-}
