--- src/by/it/VeronicaDemesh/lesson02/greedyKnapsack.txt
@@ -1,5 +1,0 @@
-4 60
-60 20
-100 50
-120 30
-100 50
