--- src/by/it/VeronicaDemesh/lesson03/A_Huffman.java
@@ -1,181 +1,0 @@
-package by.it.VeronicaDemesh.lesson03;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.util.*;
-
-//Lesson 3. A_Huffman.
-//Разработайте метод encode(File file) для кодирования строки (код Хаффмана)
-
-// По данным файла (непустой строке ss длины не более 104104),
-// состоящей из строчных букв латинского алфавита,
-// постройте оптимальный по суммарной длине беспрефиксный код.
-
-// Используйте Алгоритм Хаффмана — жадный алгоритм оптимального
-// безпрефиксного кодирования алфавита с минимальной избыточностью.
-
-// В первой строке выведите количество различных букв kk,
-// встречающихся в строке, и размер получившейся закодированной строки.
-// В следующих kk строках запишите коды букв в формате "letter: code".
-// В последней строке выведите закодированную строку. Примеры ниже
-
-//        Sample Input 1:
-//        a
-//
-//        Sample Output 1:
-//        1 1
-//        a: 0
-//        0
-
-//        Sample Input 2:
-//        abacabad
-//
-//        Sample Output 2:
-//        4 14
-//        a: 0
-//        b: 10
-//        c: 110
-//        d: 111
-//        01001100100111
-
-public class A_Huffman {
-
-    //Изучите классы Node InternalNode LeafNode
-    abstract class Node implements Comparable<Node> {
-        //абстрактный класс элемент дерева
-        //(сделан abstract, чтобы нельзя было использовать его напрямую)
-        //а только через его версии InternalNode и LeafNode
-        private final int frequence; //частота символов
-
-        //генерация кодов (вызывается на корневом узле
-        //один раз в конце, т.е. после построения дерева)
-        abstract void fillCodes(String code);
-
-        //конструктор по умолчанию
-        private Node(int frequence) {
-            this.frequence = frequence;
-        }
-
-        //метод нужен для корректной работы узла в приоритетной очереди
-        //или для сортировок
-        @Override
-        public int compareTo(Node o) {
-            return Integer.compare(frequence, o.frequence);
-        }
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////////
-    //расширение базового класса до внутреннего узла дерева
-    private class InternalNode extends Node {
-        //внутренный узел дерева
-        Node left;  //левый ребенок бинарного дерева
-        Node right; //правый ребенок бинарного дерева
-
-        //для этого дерева не существует внутренних узлов без обоих детей
-        //поэтому вот такого конструктора будет достаточно
-        InternalNode(Node left, Node right) {
-            super(left.frequence + right.frequence);
-            this.left = left;
-            this.right = right;
-        }
-
-        @Override
-        void fillCodes(String code) {
-            left.fillCodes(code + "0");
-            right.fillCodes(code + "1");
-        }
-
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////////
-    //расширение базового класса до листа дерева
-    private class LeafNode extends Node {
-        //лист
-        char symbol; //символы хранятся только в листах
-
-        LeafNode(int frequence, char symbol) {
-            super(frequence);
-            this.symbol = symbol;
-        }
-
-        @Override
-        void fillCodes(String code) {
-            //добрались до листа, значит рекурсия закончена, код уже готов
-            //и можно запомнить его в индексе для поиска кода по символу.
-            codes.put(this.symbol, code);
-        }
-    }
-
-    //индекс данных из листьев
-    static private Map<Character, String> codes = new TreeMap<>();
-
-
-    //!!!!!!!!!!!!!!!!!!!!!!!!!     НАЧАЛО ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
-    String encode(File file) throws FileNotFoundException {
-        //прочитаем строку для кодирования из тестового файла
-        Scanner scanner = new Scanner(file);
-        String s = scanner.next();
-
-        //все комментарии от тестового решения были оставлены т.к. это задание A.
-        //если они вам мешают их можно удалить
-
-        Map<Character, Integer> count = new HashMap<>();
-        //1. переберем все символы по очереди и рассчитаем их частоту в Map count
-            //для каждого символа добавим 1 если его в карте еще нет или инкремент если есть.
-        for (int i = 0; i < s.length(); i++) {
-            char letter = s.charAt(i);
-            int iter = 1;
-            if(count.containsKey(letter)){
-                iter = count.get(letter) + 1;
-            }
-            count.put(letter, iter);
-        }
-        //2. перенесем все символы в приоритетную очередь в виде листьев
-        PriorityQueue<Node> priorityQueue = new PriorityQueue<>();
-        for (Map.Entry<Character, Integer> pair : count.entrySet()) {
-            LeafNode leafNode = new LeafNode(pair.getValue(), pair.getKey());
-            priorityQueue.add(leafNode);
-        }
-        //3. вынимая по два узла из очереди (для сборки родителя)
-        //и возвращая этого родителя обратно в очередь
-        //построим дерево кодирования Хаффмана.
-        //У родителя частоты детей складываются.
-        while (priorityQueue.size() > 1) {
-            Node leftChild = priorityQueue.poll();
-            Node rightChild = priorityQueue.poll();
-            InternalNode internalNode = new InternalNode(leftChild, rightChild);
-            priorityQueue.add(internalNode);
-        }
-        priorityQueue.poll().fillCodes("");
-
-        //4. последний из родителей будет корнем этого дерева
-        //это будет последний и единственный элемент оставшийся в очереди priorityQueue.
-        StringBuilder sb = new StringBuilder();
-        //.....
-        for (int i = 0; i < s.length(); i++) {
-            char letter = s.charAt(i);
-            sb.append(codes.get(letter));
-        }
-
-        return sb.toString();
-        //01001100100111
-        //01001100100111
-    }
-    //!!!!!!!!!!!!!!!!!!!!!!!!!     КОНЕЦ ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
-
-
-    public static void main(String[] args) throws FileNotFoundException {
-        String root = System.getProperty("user.dir") + "/src/";
-        File f = new File(root + "by/it/VeronicaDemesh/lesson03/dataHuffman.txt");
-        A_Huffman instance = new A_Huffman();
-        long startTime = System.currentTimeMillis();
-        String result = instance.encode(f);
-        long finishTime = System.currentTimeMillis();
-        System.out.printf("%d %d\n", codes.size(), result.length());
-        for (Map.Entry<Character, String> entry : codes.entrySet()) {
-            System.out.printf("%s: %s\n", entry.getKey(), entry.getValue());
-        }
-        System.out.println(result);
-    }
-
-}
