--- src/by/it/VeronicaDemesh/lesson03/encodeHuffman.txt
@@ -1,6 +1,0 @@
-4 14
-a: 0
-b: 10
-c: 110
-d: 111
-01001100100111
