--- src/by/it/VeronicaDemesh/lesson03/heapData.txt
@@ -1,7 +1,0 @@
-6
-Insert 200
-Insert 10
-ExtractMax
-Insert 5
-Insert 500
-ExtractMax
