--- src/by/it/VeronicaDemesh/lesson04/dataA.txt
@@ -1,2 +1,0 @@
-5 1 5 8 12 13
-5 8 1 23 1 11
