--- src/by/it/VeronicaDemesh/lesson05/dataB.txt
@@ -1,2 +1,0 @@
-5
-2 3 9 2 9
