--- src/by/it/VeronicaDemesh/lesson05/dataC.txt
@@ -1,4 +1,0 @@
-2 3
-0 5
-7 10
-1 6 11
