+++ src/by/it/VeronicaDemesh_old/lesson02/greedyKnapsack.txt
@@ -1,0 +1,5 @@
+4 60
+60 20
+100 50
+120 30
+100 50
