+++ src/by/it/VeronicaDemesh_old/lesson03/heapData.txt
@@ -1,0 +1,7 @@
+6
+Insert 200
+Insert 10
+ExtractMax
+Insert 5
+Insert 500
+ExtractMax
