+++ src/by/it/VeronicaDemesh_old/lesson04/Lesson4Test.java
@@ -1,0 +1,62 @@
+package by.it.VeronicaDemesh_old.lesson04;
+
+import org.junit.Test;
+
+
+import java.io.FileInputStream;
+import java.io.InputStream;
+import java.util.Arrays;
+
+import static org.junit.Assert.assertTrue;
+
+public class Lesson4Test {
+    @Test
+    public void A() throws Exception {
+        String root = System.getProperty("user.dir") + "/src/";
+        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh_old/lesson04/dataA.txt");
+        A_BinaryFind instance = new A_BinaryFind();
+        //long startTime = System.currentTimeMillis();
+        int[] result=instance.findIndex(stream);
+        //long finishTime = System.currentTimeMillis();
+        StringBuilder sb=new StringBuilder();
+        for (int index:result){
+            sb.append(index).append(" ");
+        }
+        boolean ok=sb.toString().trim().equals("3 1 -1 1 -1");
+        assertTrue("A failed", ok);
+    }
+
+
+    @Test
+    public void B() throws Exception {
+        String root = System.getProperty("user.dir") + "/src/";
+        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh_old/lesson04/dataB.txt");
+        B_MergeSort instance = new B_MergeSort();
+        //long startTime = System.currentTimeMillis();
+        int[] result=instance.getMergeSort(stream);
+        //long finishTime = System.currentTimeMillis();
+        boolean ok=result.length>3;
+        int test[]=new int[result.length];
+        System.arraycopy(result,0,test,0,result.length);
+        Arrays.sort(test);
+        for (int i = 0; i < result.length; i++) {
+            ok=ok && (result[i]==test[i]);
+        }
+        assertTrue("B failed", ok);
+    }
+
+
+    @Test
+    public void C() throws Exception {
+        String root = System.getProperty("user.dir") + "/src/";
+        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh_old/lesson04/dataC.txt");
+        C_GetInversions instance = new C_GetInversions();
+        //long startTime = System.currentTimeMillis();
+        int result = instance.calc(stream);
+        //long finishTime = System.currentTimeMillis();
+        boolean ok=(2==result);
+        assertTrue("C failed", ok);
+
+    }
+
+}
