+++ src/by/it/VeronicaDemesh_old/lesson04/dataA.txt
@@ -1,0 +1,2 @@
+5 1 5 8 12 13
+5 8 1 23 1 11
