+++ src/by/it/VeronicaDemesh_old/lesson04/dataC.txt
@@ -1,0 +1,2 @@
+5
+2 3 9 2 9
