+++ src/by/it/VeronicaDemesh_old/lesson05/B_CountSort.java
@@ -1,0 +1,77 @@
+package by.it.VeronicaDemesh_old.lesson05;
+
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.InputStream;
+import java.util.Arrays;
+import java.util.Scanner;
+
+/*
+Первая строка содержит число 1<=n<=10000, вторая - n натуральных чисел, не превышающих 10.
+Выведите упорядоченную по неубыванию последовательность этих чисел.
+
+При сортировке реализуйте метод со сложностью O(n)
+
+Пример: https://karussell.wordpress.com/2010/03/01/fast-integer-sorting-algorithm-on/
+Вольный перевод: http://programador.ru/sorting-positive-int-linear-time/
+*/
+
+public class B_CountSort {
+
+
+    int[] countSort(InputStream stream) throws FileNotFoundException {
+        //подготовка к чтению данных
+        Scanner scanner = new Scanner(stream);
+        //!!!!!!!!!!!!!!!!!!!!!!!!!     НАЧАЛО ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
+        //размер массива
+        int n = scanner.nextInt();
+        int[] points = new int[n];
+
+        //читаем точки
+        for (int i = 0; i < n; i++) {
+            points[i]=scanner.nextInt();
+        }
+        //тут реализуйте логику задачи с применением сортировки подсчетом
+
+        countingSort(points);
+        //!!!!!!!!!!!!!!!!!!!!!!!!!     КОНЕЦ ЗАДАЧИ     !!!!!!!!!!!!!!!!!!!!!!!!!
+        return points;
+    }
+
+    private int[] countingSort (int[]points){
+        int low = points[0];
+        int high = 0;
+        for (int i = 0; i < points.length; i++) {
+            if (points[i] > high){
+                high = points[i];
+            }else if (points[i] < low){
+                low = points[i];
+            }
+        }
+       int[] counts = new int[high - low + 1];
+
+       for (int i = 0; i < points.length; i++) {
+           int x = points[i];
+           counts[x - low]++;
+       }
+       int pointsIndex = 0;
+       for (int i = 0; i < counts.length; i++) {
+           if (counts[i]>0){
+               Arrays.fill(points, pointsIndex, pointsIndex + counts[i], i + low);
+               pointsIndex = pointsIndex + counts[i];
+           }
+       }
+       return points;
+    }
+
+    public static void main(String[] args) throws FileNotFoundException {
+        String root = System.getProperty("user.dir") + "/src/";
+        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh_old/lesson05/dataB.txt");
+        B_CountSort instance = new B_CountSort();
+        int[] result=instance.countSort(stream);
+        for (int index:result){
+            System.out.print(index+" ");
+        }
+    }
+
+}
