+++ src/by/it/VeronicaDemesh_old/lesson05/dataA.txt
@@ -1,0 +1,4 @@
+2 3
+0 5
+7 10
+1 6 11
