+++ src/by/it/VeronicaDemesh_old/lesson05/dataB.txt
@@ -1,0 +1,2 @@
+5
+2 3 9 2 9
