+++ src/by/it/veronicaDemesh/lesson04/C_GetInversions.java
@@ -64,7 +64,7 @@ public class C_GetInversions {
 
     public static void main(String[] args) throws FileNotFoundException {
         String root = System.getProperty("user.dir") + "/src/";
-        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh/lesson04/dataC.txt");
+        InputStream stream = new FileInputStream(root + "by/it/VeronicaDemesh_old/lesson04/dataC.txt");
         C_GetInversions instance = new C_GetInversions();
         //long startTime = System.currentTimeMillis();
         int result = instance.calc(stream);
